Move pySPFM into workflows folder
