+++ setup.cfg
@@ -58,7 +58,7 @@ all =
 
 [options.entry_points]
 console_scripts =
-    pySPFM = pySPFM.pySPFM:_main
+    pySPFM = pySPFM.workflows.pySPFM:_main
 
 [options.package_data]
 * =
